backend/requirements.txt
fastapiuvicornpython-multipartgoogle-generativeaipydanticpython-dotenv